tokenisation: separated from modelling
