+++ tests/modelling/test_tokenizer.py
@@ -2,37 +2,37 @@
 import unittest
 
 # My Packages and Modules
-from hlm12nli.modelling.tokenizer import Hlm12NliTokenizer, Hlm12NliTokenizerBatchRequiredError
+from hlm12nli.modelling.tokeniser import Hlm12NliTokeniser, Hlm12NliTokeniserBatchRequiredError
 
 
-class TestHlm12NliTokenizer(unittest.TestCase):
+class TestHlm12NliTokeniser(unittest.TestCase):
     def setUp(self):
         self.vocab = ["[PAD]", "[OOV]", "[STR]", "[END]", "A", "test", "sentence", "##.", "Hudson", "##'s"]
-        self.tokenizer = Hlm12NliTokenizer(vocab=self.vocab, do_lowercase=False)
+        self.tokeniser = Hlm12NliTokeniser(vocab=self.vocab, do_lowercase=False)
 
     def test_ctor_accepts_vocab_as_dict_or_list(self):
-        t1 = Hlm12NliTokenizer(vocab=self.vocab, do_lowercase=False)
-        t2 = Hlm12NliTokenizer(vocab={token: i for i, token in enumerate(self.vocab)}, do_lowercase=False)
+        t1 = Hlm12NliTokeniser(vocab=self.vocab, do_lowercase=False)
+        t2 = Hlm12NliTokeniser(vocab={token: i for i, token in enumerate(self.vocab)}, do_lowercase=False)
         self.assertEqual(t1.vocab, t2.vocab)
 
     def test_tokenize_requires_batch(self):
-        with self.assertRaises(Hlm12NliTokenizerBatchRequiredError):
-            self.tokenizer.tokenize("A test sentence")
+        with self.assertRaises(Hlm12NliTokeniserBatchRequiredError):
+            self.tokeniser.tokenize("A test sentence")
 
     def test_tokenize_lowercases(self):
         vocab = [t.lower() if not t.startswith("[") else t for t in self.vocab]
-        tokenizer = Hlm12NliTokenizer(vocab=vocab, do_lowercase=True)
+        tokeniser = Hlm12NliTokeniser(vocab=vocab, do_lowercase=True)
         test_sentence = "A test sentence"
         expected_output = ["[STR]", "a", "test", "sentence", "[END]"]
-        self.assertListEqual(tokenizer.tokenize([test_sentence])[0], expected_output)
+        self.assertListEqual(tokeniser.tokenize([test_sentence])[0], expected_output)
 
     def test_tokenize_includes_start(self):
         test_sentence = "A test sentence"
-        self.assertEqual(self.tokenizer.tokenize([test_sentence])[0][0], self.tokenizer.token_str)
+        self.assertEqual(self.tokeniser.tokenize([test_sentence])[0][0], self.tokeniser.token_str)
 
     def test_tokenize_includes_end(self):
         test_sentence = "A test sentence"
-        self.assertEqual(self.tokenizer.tokenize([test_sentence])[0][-1], self.tokenizer.token_end)
+        self.assertEqual(self.tokeniser.tokenize([test_sentence])[0][-1], self.tokeniser.token_end)
 
     def test_tokenize_includes_pad(self):
         test_sentence = ["A test sentence", "Hudson test"]
@@ -40,20 +40,27 @@ class TestHlm12NliTokenizer(unittest.TestCase):
             ["[STR]", "A", "test", "sentence", "[END]"],
             ["[STR]", "Hudson", "test", "[END]", "[PAD]"],
         ]
-        actual_output = self.tokenizer.tokenize(test_sentence)
+        actual_output = self.tokeniser.tokenize(test_sentence)
         self.assertListEqual(actual_output, expected_output)
 
     def test_tokenize_with_only_whitespace_splitters(self):
         test_sentence = "A test sentence"
         expected_output = ["[STR]", "A", "test", "sentence", "[END]"]
-        self.assertEqual(self.tokenizer.tokenize([test_sentence])[0], expected_output)
+        self.assertEqual(self.tokeniser.tokenize([test_sentence])[0], expected_output)
 
     def test_tokenize_with_subwords(self):
         test_sentence = "Hudson's test"
         expected_output = ["[STR]", "Hudson", "##'s", "test", "[END]"]
-        self.assertEqual(self.tokenizer.tokenize([test_sentence])[0], expected_output)
+        self.assertEqual(self.tokeniser.tokenize([test_sentence])[0], expected_output)
 
     def test_tokenize_with_unk_token(self):
         test_sentence = "A test sentence not_in_vocab"
         expected_output = ["[STR]", "A", "test", "sentence", "[OOV]", "[END]"]
-        self.assertEqual(self.tokenizer.tokenize([test_sentence])[0], expected_output)
+        self.assertEqual(self.tokeniser.tokenize([test_sentence])[0], expected_output)
+
+    def test_tokenize_and_join_without_special_tokens_match(self):
+        test_sentence = ["A Hudson's test sentence."]
+        expected_output = ["A Hudson's test sentence."]
+        batch_tokenized = self.tokeniser.tokenize([test_sentence])
+        actual_output = self.tokeniser.untokenize(batch_tokenized)
+        self.assertEqual(actual_output, expected_output)
